TINKERPOP-1562 ImportCustomizer has become an interface
This is a breaking change unfortunately, but this was not a feature that was published or used in 3.2.3 so I don't expect wide usage.
