+++ gremlin-core/src/main/java/org/apache/tinkerpop/gremlin/jsr223/CoreGremlinPlugin.java
@@ -31,7 +31,7 @@ public final class CoreGremlinPlugin implements GremlinPlugin {
 
     private static final String MODULE_NAME = "tinkerpop.core";
 
-    private static final ImportCustomizer gremlinCore = ImportCustomizer.build()
+    private static final ImportCustomizer gremlinCore = DefaultImportCustomizer.build()
             .addClassImports(CoreImports.getClassImports())
             .addEnumImports(CoreImports.getEnumImports())
             .addMethodImports(CoreImports.getMethodImports()).create();
